requirements.txt
aenum==2.2.6
annotated-types==0.6.0
argcomplete==3.1.2
attrs==23.1.0
black==23.10.1
blinker==1.6.3
bunq-sdk==1.14.18
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
datamodel-code-generator==0.22.1
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
genson==1.2.2
gunicorn==21.2.0
idna==3.4
inflect==5.6.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.6
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.7
packaging==23.2
pathable==0.4.3
pathspec==0.11.2
platformdirs==3.11.0
prance==23.6.21.0
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyrsistent==0.20.0
PySnooper==1.2.0
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.10.6
ruamel.yaml==0.18.2
ruamel.yaml.clib==0.2.8
simplejson==3.19.1
six==1.16.0
termcolor==2.3.0
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.17
watchdog==3.0.0
Werkzeug==3.0.0
